Use Go's new fuzz tooling instead of go-fuzz
Use Go's new fuzz tooling instead of go-fuzz
