+++ internal/proto/testdata/fuzz/FuzzChannelData/c1c1d815dfbe98c11d89ee9fb0436ede58f3451a3a5338cd4e0f6b1122fd600f
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("*")
